requirements.txt
streamlit
google-generativeai
gtts